add bookmark python bindings
Summary: Add edenapi client python bindings for bookmarks so we use the bookmark api in python mercurial. At this layer there's a mixture of methods that iterate through the response stream and await the stats future and methods that directly return the response stream and stat future. I wasn't certain what to do in this case. The former may be cleaner later on, but the latter may be desirable if we paginate bookmark --list-remote output and therefore don't end up consuming the entire stream.

Reviewed By: kulshrax

Differential Revision: D27331446

fbshipit-source-id: 4dcbcb4bbd69117e02f2028e527f3d6f91c9bf99