mononoke: fixup some test flakyness
Summary:
Like it says in the title. This test wants to see consistent bookmarks so it
should flush them.

Note that this used to just opt out of bookmarks caching entirely, but I'd like
us to try and avoid having so many snowflakes in our tests (because it makes
their maintenance harder), so instead of changing the environment, let's change
the test to do what it needs to do.

Reviewed By: mzr, HarveyHunt

Differential Revision: D27964099

fbshipit-source-id: 72e00bad07dec15f18faaf4aa2e32e78cb333ab0