mononoke: add fastlog to warm bookmark cache
Summary:
It wasn't in warm bookmark cache, but it was an oversight - there's no reason
for it to not be here. Let's add it, since in case of crashlooping derived data
tailer (see attached task T89911396) there might be nothing to derive fastlog
data structure, and we end up with a long queue to derive.

Reviewed By: krallin

Differential Revision: D28114533

fbshipit-source-id: feb29c07d90be6250c5385ae9f2fb13eb52eedba