mononoke: enable level filtering when printing information about stdlog
Summary:
Right now we write straight to a logger with no filter, so no matter the log
level we print this stuff out. Let's fix it.

While we're at it, move this back to debug level.

I'd made this trace in my recent cmdlib refactoring (which resulted in us
properly initializing logging in all binaries), since I assumed we just had level
filtering working but with debug-logging enabled and I didn't want to have to
update every single test, but it turns out that the reason we didn't print it
out at trace is just because thats not enabled at all in our slog build:
D28097080.

Reviewed By: StanislavGlebik

Differential Revision: D28116053

fbshipit-source-id: f59d9a70ea3c3d834adea16f2686bfc244672b14