mononoke: walk from BonsaiHgMapping to HgChangesetViaBonsai for consistency
Summary:
Previously there were two different paths to HgChangeset.  This diff unifies them, so that when walker state.rs is checking for a previous visit it will find that it happened.

For existing walks of changesets in the NewestFirst direction this wasn't causing a problem, however the next diff in stack adds support for OldestFirst walks.  In the OldestFirst case the mismatch in paths to HgChangeset was leaving a deferred edge to visit when everything should have been visited in previous chunks.

Differential Revision: D28095569

fbshipit-source-id: ccba4a679fc28bde042cfc222e5097c84fa968c0