mononoke: make pack hash order independent
Summary:
Our packs no longer require strict order on write to be readable, so let's hash the list of sorted keys as the pack identity. i.e. pack will have same identity if if contains same keys, regardless of order.

This makes the packer easier to work with in general

Reviewed By: farnz

Differential Revision: D28119423

fbshipit-source-id: 71c8d4ed351bde63065047c706311a869b2beeda