mononoke_types: add definition for blame_v2 thrift types
Summary:
Add thrift definitions for blame_v2.  This uses a map from integer id to
changeset id to avoid repeating 32-byte hashes in files with many blame ranges,
and to provide a stable integer index for changesets that can be used to
provide a more human-readable id for each blamed changeset.

Reviewed By: StanislavGlebik

Differential Revision: D28183163

fbshipit-source-id: a6ae01d92c3d3c022a1d412f2ba0046f08c2d9f1