mononoke: sample harder in the walker
Summary:
The walker does about 50% of our traffic, but it also has fairly predictable
access patterns. It seems unlikely that we really benefit from logging all
scrub activity with the same precision as we do other traffic.

So, let's sample it. This should make a lot of space in our Scuba table and
make us more resilient to sudden changes in activity.

Reviewed By: StanislavGlebik

Differential Revision: D28254057

fbshipit-source-id: da748a565954c31c2d9e087b7b07747a435427bf